requirements.txt
# Web Framework
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5

# Database
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
pymongo==4.5.0

# Environment & Config
python-dotenv==1.0.0
python-decouple==3.8

# HTTP & Requests
requests==2.31.0
httpx==0.24.1

# AI & ML
openai==0.28.1
langchain==0.0.310
numpy==1.24.3
pandas==1.5.3

# Authentication & Security
PyJWT==2.8.0
werkzeug==2.3.7
cryptography==41.0.3

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Production
gunicorn==21.2.0
python-multipart==0.0.6

# Logging & Monitoring
python-json-logger==2.0.7

# Utils
click==8.1.7
